Reorganized to separate GUI, network communications, and shape drawing code to make it easier to separate the concerns.
