+++ .classpath
@@ -11,5 +11,9 @@
 	<classpathentry kind="var" path="USERLIBS_DIR/CTRE_Phoenix-javadoc.jar"/>
 	<classpathentry kind="var" path="USERLIBS_DIR/CTRE_Phoenix-sources.jar"/>
 	<classpathentry kind="var" path="USERLIBS_DIR/CTRE_Phoenix.jar"/>
+	<classpathentry kind="var" path="USERLIBS_DIR/ntcore-jni-4.0.0-all.jar"/>
+	<classpathentry kind="var" path="USERLIBS_DIR/opencv-natives-3.1.0-linux-x86_64.jar"/>
+	<classpathentry kind="var" path="USERLIBS_DIR/opencv-natives-3.1.0-osx-x86_64.jar"/>
+	<classpathentry kind="var" path="USERLIBS_DIR/opencv-natives-3.1.0-windows-x86_64_2015.jar"/>
 	<classpathentry kind="output" path="bin"/>
 </classpath>
